improve folder structure: separate backend and frontend code
