+++ frontend/components/common/AuthButton.tsx
