+++ frontend/components/listing/ListingCard.tsx
